+++ src/include/sequence.hpp
@@ -17,6 +17,8 @@ class Sequence : public std::vector<Number> {
 
   bool is_linear(size_t start) const;
 
+  int64_t get_first_non_decreasing_term() const;
+
   bool align(const Sequence &s, int64_t max_offset);
 
   bool operator==(const Sequence &s) const;