+++ src/oeis_manager.cpp
@@ -575,10 +575,27 @@ std::string OeisManager::isOptimizedBetter(Program existing, Program optimized,
     return "";
   }
 
-  // compare number of successfully computed terms
+  // if the programs are the same, no need to evaluate them
+  if (optimized == existing) {
+    return "";
+  }
+
+  // get extended sequence
+  auto terms = seq.getTerms(OeisSequence::EXTENDED_SEQ_LENGTH);
+  if (terms.empty()) {
+    Log::get().error("Error fetching b-file for " + seq.id_str(), true);
+  }
+
+  // check if the first non-decreasing term is beyond the know sequence terms
   static const int64_t num_terms = 10000;  // magic number
   Sequence tmp;
   const auto optimized_steps = evaluator.eval(optimized, tmp, num_terms, false);
+  if (tmp.get_first_non_decreasing_term() >=
+      static_cast<int64_t>(terms.size())) {
+    return "";
+  }
+
+  // compare number of successfully computed terms
   const auto existing_steps = evaluator.eval(existing, tmp, num_terms, false);
   if (optimized_steps.runs > existing_steps.runs) {
     return "Better";